requirements.txt
matrix-nio[e2e]>=0.21.0
anthropic>=0.40.0
apscheduler>=3.10.0
pyyaml>=6.0
python-dateutil>=2.9.0
aiosqlite>=0.20.0
